refactor(CISource): place CISource implementation under providers
