+++ source/ci_source/providers/index.js
@@ -1,0 +1,7 @@
+import Travis from "./Travis"
+import Circle from "./Circle"
+import Semaphore from "./Semaphore"
+import Jenkins from "./Jenkins"
+import Fake from "./Fake"
+
+export default [Travis, Circle, Semaphore, Jenkins, Fake]
